+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [UNRELEASED]
+
+- added custom commands availability to all modes
+
 ## [0.21.0]
 
 - not storing tool messages without the usage report into DB